* There are no more IDs than the possible number of indices. This plugs   compiler warnings.
git-svn-id: https://ecell.svn.sourceforge.net/svnroot/ecell/ecell3/trunk@2750 f1531174-cb10-0410-9fe6-89aa7ac3eedb

+++ ecell/libecs/DynamicPriorityQueue.hpp
@@ -57,10 +57,10 @@ class PersistentIDPolicy
 
 public:
 
-    typedef unsigned long long ID;
-
+    typedef size_t                 ID;
     typedef std::vector< ID >      IDVector;
     typedef IDVector::size_type    Index;
+
 
 #if defined( HAVE_UNORDERED_MAP ) || defined( HAVE_TR1_UNORDERED_MAP )
